+++ src/main/java/coordinate/domain/Figure/FigureFactory.java
@@ -2,7 +2,20 @@ package coordinate.domain.Figure;
 
 import coordinate.domain.point.PointGroup;
 
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Function;
+
 public class FigureFactory implements FigureCreator {
+    Map<Integer, Function<PointGroup, Figure>> figures = new HashMap<Integer, Function<PointGroup, Figure>>() {
+        {
+            put(1, (points) -> null);
+            put(2, (points) -> new Line(points));
+            put(3, (points) -> new Triangle(points));
+            put(4, (points) -> new Rectangle(points));
+        }
+    };
+
     @Override
     public Figure create(PointGroup points) {
         try {
