+++ src/main/java/rentcar/Reportable.java
@@ -1,7 +1,7 @@
 package rentcar;
 
 interface Reportable {
-    String NEWLINE = System.getProperty("line.separator");
+    String NEWLINE = "\n";
 
     String generateReport();
 }
